Moved event status and provide additional clean up for double items. - extra controls added to team randomization based on skills with extra item added to the dialog controls. -

+++ app/src/main/java/com/example/seanrafferty/mygrouprandomiser/AddPlayerActivity.kt
@@ -73,7 +73,7 @@ class AddPlayerActivity : AppCompatActivity() {
         if(name.isNotEmpty())
         {
             var playerManager = PlayerManager(this)
-            return playerManager.SavePlayer(Player(0, name, rating.toString().toInt()))
+            return playerManager.SavePlayer(Player(0, name, rating.toString().toDouble()))
         }
         Toast.makeText(this, "No Name Provided", Toast.LENGTH_LONG).show()
         return 0
